Add opt-in support for functions
This addresses #10 and causes the least impact to existing users
and OpenFaaS deployments by making faas-idler an opt-in rather
than opt-out decision.

Functions require a new label and this is documented on the
README. Tested with Swarm by deploying three functions - one with
no label, one with a label with "false" as the scale value and one
with "true" set.

If it makes sense to change tact later on we can do this in the
faas-idler configuration options.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>